created method & exception for retrieving environmental variables
