+++ src/main/java/com/codecool/testautomation/norsemanonsalad/testutils/EnvironmentalVariableNotFoundException.java
@@ -1,0 +1,7 @@
+package com.codecool.testautomation.norsemanonsalad.testutils;
+
+public class EnvironmentalVariableNotFoundException extends RuntimeException {
+    public EnvironmentalVariableNotFoundException(String message) {
+        super(message);
+    }
+}
