+++ src/test/java/com/codecool/testautomation/norsemanonsalad/testutils/ExcelReaderTest.java
@@ -1,4 +1,4 @@
-package com.codecool.testautomation.NorseManonSalad.TestUtils;
+package com.codecool.testautomation.NorseManonSalad.testutils;
 
 import org.junit.jupiter.api.Test;
